refactor(dev-core): consolidate shims into adapters, fix cross-plugin boundary
Closes #25

V1: Move config.ts and workspace.ts unique exports into
adapters/config-helpers.ts and ports/workspace.ts respectively.
Update 22 import sites. Delete both shim files.

V2: Move github.ts standalone functions into adapters/github-adapter.ts.
Update 11 import sites + 8 test vi.mock paths. Delete shim file.

V3: Add vault_index_entry() to roxabi_sdk/paths.py. Update
get-invoice-details/scripts/extract.py to use SDK function instead of
raw sqlite3, fixing cross-plugin boundary violation.

All 9 success criteria verified:
- 3 shim files deleted
- Zero old import paths remaining
- vault_index_entry() available in SDK
- bun typecheck + lint + test pass (204 tests, 0 failures)
