+++ plugins/dev-core/skills/init/__tests__/protection.test.ts
@@ -1,6 +1,6 @@
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest'
 
-vi.mock('../../shared/config', () => ({
+vi.mock('../../shared/adapters/config-helpers', () => ({
   PROTECTED_BRANCHES: ['main', 'staging'],
   BRANCH_PROTECTION_PAYLOAD: {
     required_status_checks: { strict: true, contexts: ['ci'] },
@@ -9,7 +9,7 @@ vi.mock('../../shared/config', () => ({
   },
 }))
 
-vi.mock('../../shared/github', () => ({
+vi.mock('../../shared/adapters/github-adapter', () => ({
   run: vi.fn(),
 }))
 
@@ -20,7 +20,7 @@ describe('protectBranches', () => {
 
   beforeEach(async () => {
     vi.restoreAllMocks()
-    const github = await import('../../shared/github')
+    const github = await import('../../shared/adapters/github-adapter')
     mockRun = github.run as ReturnType<typeof vi.fn>
     mockRun.mockResolvedValue('')
