+++ plugins/dev-core/skills/init/lib/discover.ts
@@ -3,8 +3,8 @@
  * Returns structured JSON for the init orchestrator.
  */
 
-import { PROTECTED_BRANCHES, STANDARD_LABELS, STANDARD_WORKFLOWS } from '../../shared/config'
-import { getBoardIssueNumbers, parseProjectFields, run } from '../../shared/github'
+import { PROTECTED_BRANCHES, STANDARD_LABELS, STANDARD_WORKFLOWS } from '../../shared/adapters/config-helpers'
+import { getBoardIssueNumbers, parseProjectFields, run } from '../../shared/adapters/github-adapter'
 import { checkPrereqs } from '../../shared/prereqs'
 
 export interface DiscoveryResult {
