+++ plugins/dev-core/skills/init/lib/labels.ts
@@ -2,8 +2,8 @@
  * Create standard dev-core labels on the GitHub repo.
  */
 
-import { type LabelDef, STANDARD_LABELS } from '../../shared/config'
-import { run } from '../../shared/github'
+import { type LabelDef, STANDARD_LABELS } from '../../shared/adapters/config-helpers'
+import { run } from '../../shared/adapters/github-adapter'
 
 export type LabelScope = 'all' | 'type' | 'area' | 'priority'
