+++ plugins/dev-core/skills/issue-triage/__tests__/create.test.ts
@@ -30,7 +30,7 @@ process.env.PRIORITY_OPTIONS_JSON = JSON.stringify({
 
 // Mock config before github — vi.mock is hoisted before process.env assignments,
 // so importOriginal would load config.ts with empty env vars. Manual factory is required.
-vi.mock('../../shared/config', () => ({
+vi.mock('../../shared/adapters/config-helpers', () => ({
   isProjectConfigured: () => true,
   NOT_CONFIGURED_MSG: 'GitHub Project V2 is not configured.',
   GH_PROJECT_ID: 'PVT_test',
@@ -88,7 +88,7 @@ vi.mock('../../shared/config', () => ({
   },
 }))
 
-vi.mock('../../shared/github', () => ({
+vi.mock('../../shared/adapters/github-adapter', () => ({
   createGitHubIssue: vi.fn(),
   getNodeId: vi.fn(),
   addToProject: vi.fn(),
@@ -97,7 +97,7 @@ vi.mock('../../shared/github', () => ({
   addSubIssue: vi.fn(),
 }))
 
-const github = await import('../../shared/github')
+const github = await import('../../shared/adapters/github-adapter')
 const mockCreateGitHubIssue = github.createGitHubIssue as ReturnType<typeof vi.fn>
 const mockGetNodeId = github.getNodeId as ReturnType<typeof vi.fn>
 const mockAddToProject = github.addToProject as ReturnType<typeof vi.fn>
